fix(tests): use non-zero recovery timeout in circuit breaker tests
Tests using recovery_timeout_seconds=0 caused instant HALF_OPEN
transition, making the OPEN→HALF_OPEN transition untestable. Changed
to 0.5s with proper sleep to verify the timeout actually elapses.
Also widened type hint from int to float to support sub-second values.

+++ src/legion_koi/resilience/circuit_breaker.py
@@ -58,7 +58,7 @@ class CircuitBreaker:
         self,
         name: str,
         failure_threshold: int = CIRCUIT_FAILURE_THRESHOLD,
-        recovery_timeout_seconds: int = CIRCUIT_RECOVERY_TIMEOUT_SECONDS,
+        recovery_timeout_seconds: float = CIRCUIT_RECOVERY_TIMEOUT_SECONDS,
         event_bus=None,
     ):
         self.name = name
